@@ -3,20 +3,23 @@
  * Licensed under the MIT License. See License.txt in the project root for
  * license information.
  */
-package com.microsoft.azure.management.eventhub.samples;
+package com.azure.resourcemanager.eventhub.samples;
 
-import com.microsoft.azure.management.Azure;
-import com.microsoft.azure.management.cosmosdb.CosmosDBAccount;
-import com.microsoft.azure.management.cosmosdb.DatabaseAccountKind;
-import com.microsoft.azure.management.eventhub.EventHubNamespace;
-import com.microsoft.azure.management.eventhub.EventHubNamespaceAuthorizationRule;
-import com.microsoft.azure.management.monitor.DiagnosticSetting;
-import com.microsoft.azure.management.resources.fluentcore.arm.Region;
-import com.microsoft.azure.management.samples.Utils;
-import com.microsoft.rest.LogLevel;
-import org.joda.time.Period;
+import java.time.Duration;
 
-import java.io.File;
+import com.azure.core.credential.TokenCredential;
+import com.azure.core.http.policy.HttpLogDetailLevel;
+import com.azure.core.management.AzureEnvironment;
+import com.azure.core.management.Region;
+import com.azure.core.management.profile.AzureProfile;
+import com.azure.identity.DefaultAzureCredentialBuilder;
+import com.azure.resourcemanager.AzureResourceManager;
+import com.azure.resourcemanager.cosmos.models.CosmosDBAccount;
+import com.azure.resourcemanager.cosmos.models.DatabaseAccountKind;
+import com.azure.resourcemanager.eventhubs.models.EventHubNamespace;
+import com.azure.resourcemanager.eventhubs.models.EventHubNamespaceAuthorizationRule;
+import com.azure.resourcemanager.monitor.models.DiagnosticSetting;
+import com.azure.resourcemanager.samples.Utils;
 
 /**
  * Stream Azure Service Logs and Metrics for consumption through Event Hub.
@@ -32,12 +35,10 @@ public class ManageEventHubEvents {
      * @param azure instance of the azure client
      * @return true if sample runs successfully
      */
-    public static boolean runSample(Azure azure) {
-        final String storageAccountConnectionStringFormat = "DefaultEndpointsProtocol=https;AccountName=%s;AccountKey=%s;EndpointSuffix=core.windows.net";
+    public static boolean runSample(AzureResourceManager azureResourceManager) {
         final Region region = Region.US_EAST;
-        final String rgName = Utils.createRandomName("rgEvHb");
-        final String namespaceName = Utils.createRandomName("ns");
-        final String storageAccountName = Utils.createRandomName("stg");
+        final String rgName = Utils.randomResourceName(azureResourceManager, "rgEvHb", 14);
+        final String namespaceName = Utils.randomResourceName(azureResourceManager, "ns", 14);
         final String eventHubName = "FirstEventHub";
         String diagnosticSettingId = null;
 
@@ -45,7 +46,7 @@ public class ManageEventHubEvents {
             //=============================================================
             // Creates a Cosmos DB.
             //
-            CosmosDBAccount docDb = azure.cosmosDBAccounts()
+            CosmosDBAccount docDb = azureResourceManager.cosmosDBAccounts()
                     .define(namespaceName)
                     .withRegion(region)
                     .withNewResourceGroup(rgName)
@@ -63,7 +64,7 @@ public class ManageEventHubEvents {
 
             System.out.println("Creating event hub namespace and event hub");
 
-            EventHubNamespace namespace = azure.eventHubNamespaces()
+            EventHubNamespace namespace = azureResourceManager.eventHubNamespaces()
                     .define(namespaceName)
                     .withRegion(region)
                     .withExistingResourceGroup(rgName)
@@ -80,7 +81,7 @@ public class ManageEventHubEvents {
 
             System.out.println("Retrieving the namespace authorization rule");
 
-            EventHubNamespaceAuthorizationRule eventHubAuthRule = azure.eventHubNamespaces().authorizationRules()
+            EventHubNamespaceAuthorizationRule eventHubAuthRule = azureResourceManager.eventHubNamespaces().authorizationRules()
                     .getByName(namespace.resourceGroupName(), namespace.name(), "RootManageSharedAccessKey");
 
             System.out.println("Namespace authorization rule Retrieved");
@@ -92,13 +93,13 @@ public class ManageEventHubEvents {
             System.out.println("Enabling diagnostics events of a cosmosdb to stream to event hub");
 
             // Store Id of created Diagnostic settings only for clean-up
-            DiagnosticSetting ds  = azure.diagnosticSettings()
+            DiagnosticSetting ds  = azureResourceManager.diagnosticSettings()
                     .define("DiaEventHub")
                         .withResource(docDb.id())
                         .withEventHub(eventHubAuthRule.id(), eventHubName)
                         .withLog("DataPlaneRequests", 0)
                         .withLog("MongoRequests", 0)
-                        .withMetric("AllMetrics", Period.minutes(5), 0)
+                        .withMetric("AllMetrics", Duration.ofMinutes(5), 0)
                     .create();
 
             Utils.print(ds);
@@ -117,10 +118,10 @@ public class ManageEventHubEvents {
             try {
                 if (diagnosticSettingId != null) {
                     System.out.println("Deleting Diagnostic Setting: " + diagnosticSettingId);
-                    azure.diagnosticSettings().deleteById(diagnosticSettingId);
+                    azureResourceManager.diagnosticSettings().deleteById(diagnosticSettingId);
                 }
                 System.out.println("Deleting Resource Group: " + rgName);
-                azure.resourceGroups().beginDeleteByName(rgName);
+                azureResourceManager.resourceGroups().beginDeleteByName(rgName);
                 System.out.println("Deleted Resource Group: " + rgName);
             } catch (NullPointerException npe) {
                 System.out.println("Did not create any resources in Azure. No clean up is necessary");
@@ -142,17 +143,16 @@ public class ManageEventHubEvents {
             //=============================================================
             // Authenticate
 
-            final File credFile = new File(System.getenv("AZURE_AUTH_LOCATION"));
+            final AzureProfile profile = new AzureProfile(AzureEnvironment.AZURE);
+            final TokenCredential credential = new DefaultAzureCredentialBuilder()
+                    .authorityHost(profile.getEnvironment().getActiveDirectoryEndpoint()).build();
 
-            Azure azure = Azure.configure()
-                    .withLogLevel(LogLevel.BASIC)
-                    .authenticate(credFile)
-                    .withDefaultSubscription();
-
+            AzureResourceManager azureResourceManager = AzureResourceManager.configure()
+                    .withLogLevel(HttpLogDetailLevel.BASIC).authenticate(credential, profile).withDefaultSubscription();
             // Print selected subscription
-            System.out.println("Selected subscription: " + azure.subscriptionId());
+            System.out.println("Selected subscription: " + azureResourceManager.subscriptionId());
 
-            runSample(azure);
+            runSample(azureResourceManager);
         } catch (Exception e) {
             System.out.println(e.getMessage());
             e.printStackTrace();
